Adapt linker script for GCC ARM Embedded compiler suite
https://launchpad.net/gcc-arm-embedded

That compiler outputs a new required section: .data.rel.ro.local

Before, it was not listed in the linker script and the linker just
appended it at the end. This resulted in a .bin file that was too large
and got truncated on the Brick.

With the new section listed the .bin file size is correct again and the
firmware works.